+++ tests/conftest.py
@@ -1,9 +1,3 @@
-import json
-import os
-import sys
-from pathlib import Path
-from uuid import uuid4
-
 import pytest
 
 import substra
@@ -13,10 +7,6 @@ from . import data_factory
 from .fl_interface import FLFunctionInputs
 from .fl_interface import FLFunctionOutputs
 from .fl_interface import FunctionCategory
-from substra.tools.task_resources import TaskResources
-from substra.tools.utils import import_module
-from substra.tools.workspace import FunctionWorkspace
-from tests.tools.utils import OutputIdentifiers
 
 
 def pytest_configure(config):
@@ -141,69 +131,3 @@ def asset_factory():
 @pytest.fixture()
 def data_sample(asset_factory):
     return asset_factory.create_data_sample()
-
-
-@pytest.fixture(autouse=True)
-def patch_cwd(monkeypatch, workdir):
-    # this is needed to ensure the workspace is located in a tmpdir
-    def getcwd():
-        return str(workdir)
-
-    monkeypatch.setattr(os, "getcwd", getcwd)
-
-
-@pytest.fixture()
-def valid_opener_code():
-    return """
-import json
-from substratools import Opener
-
-class FakeOpener(Opener):
-    def get_data(self, folder):
-        return 'X', list(range(0, 3))
-
-    def fake_data(self, n_samples):
-        return ['Xfake'] * n_samples, [0] * n_samples
-"""
-
-
-@pytest.fixture()
-def valid_opener(valid_opener_code):
-    import_module("opener", valid_opener_code)
-    yield
-    del sys.modules["opener"]
-
-
-@pytest.fixture()
-def valid_opener_script(workdir, valid_opener_code):
-    opener_path = workdir / "my_opener.py"
-    opener_path.write_text(valid_opener_code)
-
-    return str(opener_path)
-
-
-@pytest.fixture(autouse=True)
-def output_model_path(workdir: Path) -> str:
-    path = workdir / str(uuid4())
-    yield path
-    if path.exists():
-        os.remove(path)
-
-
-@pytest.fixture(autouse=True)
-def output_model_path_2(workdir: Path) -> str:
-    path = workdir / str(uuid4())
-    yield path
-    if path.exists():
-        os.remove(path)
-
-
-@pytest.fixture()
-def valid_function_workspace(output_model_path: str) -> FunctionWorkspace:
-    workspace_outputs = TaskResources(
-        json.dumps([{"id": OutputIdentifiers.shared, "value": str(output_model_path), "multiple": False}])
-    )
-
-    workspace = FunctionWorkspace(outputs=workspace_outputs)
-
-    return workspace
